Update developer tooling (#2)
The developer tooling in the repository has been updated to match my
personal preferences. For one, formatters and linters have been added
for common filetypes, pre-commit hooks are used to enforce a consistent
coding style, and the `Earthfile` has been rewritten to support this new
setup.

The setup for Renovate has also been simplified, removing the helpers
and different branches that the upstream repository used.