Skip restoring cache when `use-cache` is `false` (#11)
The action has an option to skip the cache, but this has only been
implemented for saving the cache. Even when `use-cache` was set to
`false`, the action tried to to restore the binary from the cache. The
same check is now implemented for saving and restoring the cache,
ensuring that the option is respected fully.